Adding 64 bit support

+++ BulletPoint/ListViewController.h
@@ -12,12 +12,12 @@
 @interface ListViewController : UIViewController
 
 @property (nonatomic,strong) NSString* listTitle;
-@property (nonatomic,assign) int listIndex;
+@property (nonatomic,assign) long listIndex;
 @property (nonatomic,strong) MainViewController* parent;
 
 - (void)checkRow:(NSString*)title;
 - (void)uncheckRow:(NSString*)title;
-- (id)initWithIndex:(int)index;
+- (id)initWithIndex:(long)index;
 - (void)addItem:(NSString*)title;
 - (void)considerOpeningKeyboard;
 - (void)rename:(NSString*)itemID to:(NSString*)to;
